set up hierarchy for start
